+++ App.tsx
@@ -1,12 +1,16 @@
 import React from 'react'
 import { NavigationContainer } from '@react-navigation/native'
 import RootNavigator from './src /navigation/RootNavigator'
+import { Provider } from 'react-redux'
+import { store } from './src /state/store'
 
 const App = () => {
   return (
-    <NavigationContainer>
-      <RootNavigator />
-    </NavigationContainer>
+    <Provider store={store}>
+      <NavigationContainer>
+        <RootNavigator />
+      </NavigationContainer>
+    </Provider>
   )
 }