+++ src /navigation/RootNavigator.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ProductScreen from '../screens/ProductScreen';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import CartScreen from '../screens/CartScreen';
 
 const Stack = createNativeStackNavigator();
 
@@ -8,6 +9,7 @@ const RootNavigator = () => {
     return (
         <Stack.Navigator screenOptions={{ headerShown: false }} initialRouteName='Product'>
             <Stack.Screen name="Product" component={ProductScreen} />
+            <Stack.Screen name="Cart" component={CartScreen} />
         </Stack.Navigator>
     )
 }